+++ sdk/core/azure-core/inc/azure/core/context.hpp
@@ -8,6 +8,8 @@
 
 #pragma once
 
+#include "azure/core/datetime.hpp"
+
 #include <atomic>
 #include <chrono>
 #include <memory>
@@ -35,45 +37,56 @@ namespace Azure { namespace Core {
   };
 
   /**
-   * @brief A context is a node within a tree that represents expiration times and key/value pairs.
+   * @brief A context is a node within a tree that represents deadlines and key/value pairs.
    */
   class Context {
   public:
     /**
-     * @brief A type used to provide a point in time for context expiration.
+     * @brief A context key.
      */
-    using time_point = std::chrono::system_clock::time_point;
+    class Key final {
+      Key const* m_uniqueAddress;
+
+    public:
+      Key() : m_uniqueAddress(this) {}
+
+      bool operator==(Key const& other) const
+      {
+        return this->m_uniqueAddress == other.m_uniqueAddress;
+      }
+
+      bool operator!=(Key const& other) const { return !(*this == other); }
+    };
 
   private:
     struct ContextSharedState
     {
       std::shared_ptr<ContextSharedState> Parent;
-      std::atomic_int64_t CancelAtMsecSinceEpoch;
-      std::string Key;
+      std::atomic<DateTime::rep> Deadline;
+      Context::Key Key;
       std::shared_ptr<void> Value;
       const std::type_info& ValueType;
 
-      static constexpr int64_t ToMsecSinceEpoch(time_point time)
+      static constexpr DateTime::rep ToDateTimeRepresentation(DateTime const& dateTime)
       {
-        return static_cast<int64_t>(
-            std::chrono::duration_cast<std::chrono::milliseconds>(time - time_point()).count());
+        return dateTime.time_since_epoch().count();
       }
 
-      static constexpr time_point FromMsecSinceEpoch(int64_t msec)
+      static constexpr DateTime FromDateTimeRepresentation(DateTime::rep dtRepresentation)
       {
-        return time_point() + static_cast<std::chrono::milliseconds>(msec);
+        return DateTime(DateTime::time_point(DateTime::duration(dtRepresentation)));
       }
 
       explicit ContextSharedState()
-          : CancelAtMsecSinceEpoch(ToMsecSinceEpoch((time_point::max)())), Value(nullptr),
+          : Deadline(ToDateTimeRepresentation((DateTime::max)())), Value(nullptr),
             ValueType(typeid(std::nullptr_t))
       {
       }
 
       explicit ContextSharedState(
           const std::shared_ptr<ContextSharedState>& parent,
-          time_point cancelAt)
-          : Parent(parent), CancelAtMsecSinceEpoch(ToMsecSinceEpoch(cancelAt)), Value(nullptr),
+          DateTime const& deadline)
+          : Parent(parent), Deadline(ToDateTimeRepresentation(deadline)), Value(nullptr),
             ValueType(typeid(std::nullptr_t))
       {
       }
@@ -81,10 +94,10 @@ namespace Azure { namespace Core {
       template <class T>
       explicit ContextSharedState(
           const std::shared_ptr<ContextSharedState>& parent,
-          time_point cancelAt,
-          const std::string& key,
+          DateTime const& deadline,
+          Context::Key const& key,
           T value) // NOTE, should this be T&&
-          : Parent(parent), CancelAtMsecSinceEpoch(ToMsecSinceEpoch(cancelAt)), Key(key),
+          : Parent(parent), Deadline(ToDateTimeRepresentation(deadline)), Key(key),
             Value(std::make_shared<T>(std::move(value))), ValueType(typeid(T))
       {
       }
@@ -97,11 +110,9 @@ namespace Azure { namespace Core {
     {
     }
 
-    time_point CancelWhen() const;
-
   public:
     /**
-     * @brief Construct a new context with no expiration, and no value associated.
+     * @brief Construct a new context with no deadline, and no value associated.
      */
     Context() : m_contextSharedState(std::make_shared<ContextSharedState>()) {}
 
@@ -111,35 +122,43 @@ namespace Azure { namespace Core {
     Context& operator=(const Context&) = default;
 
     /**
-     * @brief Create a context with expiration.
+     * @brief Create a context with a deadline.
      *
-     * @param cancelWhen A point in time after which a context expires.
+     * @param deadline A point in time after which a context expires.
      *
-     * @return A child context with expiration.
+     * @return A child context with deadline.
      */
-    Context WithDeadline(time_point cancelWhen) const
+    Context WithDeadline(DateTime const& deadline) const
     {
-      return Context{std::make_shared<ContextSharedState>(m_contextSharedState, cancelWhen)};
+      return Context{std::make_shared<ContextSharedState>(m_contextSharedState, deadline)};
     }
 
     /**
-     * @brief Create a context without an expiration, but with \p key and \p value associated with
-     * it.
+     * @brief Create a context without a deadline, but with \p key and \p value associated with it.
      *
      * @param key A key to associate with this context.
      * @param value A value to associate with this context.
      *
-     * @return A child context with no expiration and the \p key and \p value associated with it.
+     * @return A child context with no deadline and the \p key and \p value associated with it.
      */
-    template <class T> Context WithValue(const std::string& key, T&& value) const
+    template <class T> Context WithValue(Key const& key, T&& value) const
     {
       return Context{std::make_shared<ContextSharedState>(
-          m_contextSharedState, (time_point::max)(), key, std::forward<T>(value))};
+          m_contextSharedState, (DateTime::max)(), key, std::forward<T>(value))};
     }
 
     /**
-     * @brief Try to get a value associated with a \p key parameter within this context or the
-     * branch of contexts this context belongs to.
+     * @brief Get a deadline associated with this context or the branch of contexts this context
+     * belongs to.
+     *
+     * @return A deadline associated with the context found; `Azure::DateTime::max()` value if a
+     * specific value can't be found.
+     */
+    DateTime GetDeadline() const;
+
+    /**
+     * @brief Get a value associated with a \p key parameter within this context or the branch of
+     * contexts this context belongs to.
      *
      * @param key A key associated with a context to find.
      * @param outputValue A reference to the value corresponding to the key to be set, if found
@@ -148,22 +167,18 @@ namespace Azure { namespace Core {
      * @return If found, returns true, with outputValue set to the value associated with the context
      * found; otherwise returns false.
      */
-    template <class T> bool TryGetValue(const std::string& key, T& outputValue) const
+    template <class T> const T& GetValue(Key const& key) const
     {
-      if (!key.empty())
+      for (auto ptr = m_contextSharedState; ptr; ptr = ptr->Parent)
       {
-        for (auto ptr = m_contextSharedState; ptr; ptr = ptr->Parent)
+        if (ptr->Key == key)
         {
-          if (ptr->Key == key)
+          if (typeid(T) != ptr->ValueType)
           {
-            if (typeid(T) != ptr->ValueType)
-            {
-              // type mismatch
-              std::abort();
-            }
-            outputValue = *reinterpret_cast<const T*>(ptr->Value.get());
-            return true;
+            // type mismatch
+            std::abort();
           }
+          return *reinterpret_cast<const T*>(ptr->Value.get());
         }
       }
       return false;
@@ -178,21 +193,13 @@ namespace Azure { namespace Core {
      * @return A value associated with the context found; an empty value if a specific value can't
      * be found.
      */
-    template <class T> const T& Get(const std::string& key) const
+    bool HasKey(Key const& key) const
     {
-      if (!key.empty())
+      for (auto ptr = m_contextSharedState; ptr; ptr = ptr->Parent)
       {
-        for (auto ptr = m_contextSharedState; ptr; ptr = ptr->Parent)
+        if (ptr->Key == key)
         {
-          if (ptr->Key == key)
-          {
-            if (typeid(T) != ptr->ValueType)
-            {
-              // type mismatch
-              std::abort();
-            }
-            return *reinterpret_cast<const T*>(ptr->Value.get());
-          }
+          return true;
         }
       }
       std::abort();
@@ -205,15 +212,15 @@ namespace Azure { namespace Core {
      */
     void Cancel()
     {
-      m_contextSharedState->CancelAtMsecSinceEpoch
-          = ContextSharedState::ToMsecSinceEpoch((time_point::min)());
+      m_contextSharedState->Deadline
+          = ContextSharedState::ToDateTimeRepresentation((DateTime::min)());
     }
 
     /**
      * @brief Check if the context is cancelled.
      * @return `true` if this context is cancelled, `false` otherwise.
      */
-    bool IsCancelled() const { return CancelWhen() < std::chrono::system_clock::now(); }
+    bool IsCancelled() const { return GetDeadline() < std::chrono::system_clock::now(); }
 
     /**
      * @brief Throw an exception if the context was cancelled.