+++ sdk/core/azure-core/test/ut/simplified_header.cpp
@@ -44,7 +44,6 @@ TEST(SimplifiedHeader, core)
   EXPECT_NO_THROW(Azure::Core::Uuid::CreateUuid());
   EXPECT_NO_THROW(Azure::Core::RequestFailedException("foo"));
   EXPECT_NO_THROW(Azure::Core::OperationStatus("foo"));
-  EXPECT_NO_THROW(Azure::Core::PackageVersion::VersionString());
 
   {
     std::vector<uint8_t> buffer(10);