+++ sdk/storage/azure-storage-files-datalake/src/datalake_sas_builder.cpp
@@ -124,7 +124,7 @@ namespace Azure { namespace Storage { namespace Sas {
     std::string resource = DataLakeSasResourceToString(Resource);
 
     std::string startsOnStr = StartsOn.HasValue()
-        ? StartsOn.GetValue().ToString(
+        ? StartsOn.Value().ToString(
             Azure::DateTime::DateFormat::Rfc3339, Azure::DateTime::TimeFractionFormat::Truncate)
         : "";
     std::string expiresOnStr = Identifier.empty()
@@ -133,7 +133,7 @@ namespace Azure { namespace Storage { namespace Sas {
         : "";
 
     std::string stringToSign = Permissions + "\n" + startsOnStr + "\n" + expiresOnStr + "\n"
-        + canonicalName + "\n" + Identifier + "\n" + (IPRange.HasValue() ? IPRange.GetValue() : "")
+        + canonicalName + "\n" + Identifier + "\n" + (IPRange.HasValue() ? IPRange.Value() : "")
         + "\n" + protocol + "\n" + _internal::DefaultSasVersion + "\n" + resource + "\n" + "\n"
         + CacheControl + "\n" + ContentDisposition + "\n" + ContentEncoding + "\n" + ContentLanguage
         + "\n" + ContentType;
@@ -156,7 +156,7 @@ namespace Azure { namespace Storage { namespace Sas {
     }
     if (IPRange.HasValue())
     {
-      builder.AppendQueryParameter("sip", _internal::UrlEncodeQueryParameter(IPRange.GetValue()));
+      builder.AppendQueryParameter("sip", _internal::UrlEncodeQueryParameter(IPRange.Value()));
     }
     if (!Identifier.empty())
     {
@@ -205,7 +205,7 @@ namespace Azure { namespace Storage { namespace Sas {
     std::string resource = DataLakeSasResourceToString(Resource);
 
     std::string startsOnStr = StartsOn.HasValue()
-        ? StartsOn.GetValue().ToString(
+        ? StartsOn.Value().ToString(
             Azure::DateTime::DateFormat::Rfc3339, Azure::DateTime::TimeFractionFormat::Truncate)
         : "";
     std::string expiresOnStr = ExpiresOn.ToString(
@@ -220,7 +220,7 @@ namespace Azure { namespace Storage { namespace Sas {
         + userDelegationKey.SignedTenantId + "\n" + signedStartsOnStr + "\n" + signedExpiresOnStr
         + "\n" + userDelegationKey.SignedService + "\n" + userDelegationKey.SignedVersion + "\n"
         + PreauthorizedAgentObjectId + "\n" + AgentObjectId + "\n" + CorrelationId + "\n"
-        + (IPRange.HasValue() ? IPRange.GetValue() : "") + "\n" + protocol + "\n"
+        + (IPRange.HasValue() ? IPRange.Value() : "") + "\n" + protocol + "\n"
         + _internal::DefaultSasVersion + "\n" + resource + "\n" + "\n" + CacheControl + "\n"
         + ContentDisposition + "\n" + ContentEncoding + "\n" + ContentLanguage + "\n" + ContentType;
 
@@ -240,7 +240,7 @@ namespace Azure { namespace Storage { namespace Sas {
     builder.AppendQueryParameter("sp", Permissions);
     if (IPRange.HasValue())
     {
-      builder.AppendQueryParameter("sip", _internal::UrlEncodeQueryParameter(IPRange.GetValue()));
+      builder.AppendQueryParameter("sip", _internal::UrlEncodeQueryParameter(IPRange.Value()));
     }
     builder.AppendQueryParameter("spr", _internal::UrlEncodeQueryParameter(protocol));
     builder.AppendQueryParameter(
@@ -269,7 +269,7 @@ namespace Azure { namespace Storage { namespace Sas {
     if (DirectoryDepth.HasValue())
     {
       builder.AppendQueryParameter(
-          "sdd", _internal::UrlEncodeQueryParameter(std::to_string(DirectoryDepth.GetValue())));
+          "sdd", _internal::UrlEncodeQueryParameter(std::to_string(DirectoryDepth.Value())));
     }
     if (!CacheControl.empty())
     {