+++ sdk/storage/azure-storage-common/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 12.0.0-beta.10 (Unreleased)
 
+### New Features
+
+- Added server timeout support.
+
+### Breaking Changes
+
+- Removed `Azure::Storage::Common::PackageVersion`.
 
 ## 12.0.0-beta.9 (2021-03-23)